first pass at rearchitecting and reorganizing code structure; failing commit
