--- src/components/Root.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import Navbar from './Navbar';
-import Footer from './Footer';
-
-const Root = ({ children }) => (
-  <div id="main">
-    <Navbar />
-     { children }
-    <Footer />
-  </div>
-)
-
-export default Root;
