+++ src/containers/ChartContainer.jsx
@@ -1,10 +1,9 @@
-
 import { connect } from 'react-redux';
 
-import Chart from '../components/Chart';
+import Chart from '../components/chart';
 
 const mapState = state => ({
   ...state
 });
 
-export default connect(mapState, null)( Chart );
+export default connect(mapState, null)(Chart);
