+++ src/elements/App.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import Navbar from './Navbar';
+import Footer from './Footer';
+
+const App = ({ children }) => {
+  return (
+    <div id="main">
+      <Navbar />
+      { children }
+      <Footer />
+    </div>
+  );
+}
+
+export default App;
