+++ src/elements/Navbar.jsx
@@ -4,9 +4,8 @@ import { Navbar } from 'react-materialize';
 
 const Nav = () => (
   <Navbar style={{ backgroundColor: "#2266bb" }} brand='Figure out what you need' left>
-    <li><Link to='/calculator'>Calculator</Link></li>
+    <li><Link to='/'>Calculator</Link></li>
   </Navbar>
 )
 
 export default Nav;
-
