+++ src/elements/bar-chart/index.jsx
@@ -1,9 +1,11 @@
 import React, { Component } from 'react';
-import { ResponsiveContainer, XAxis, YAxis, BarChart, Bar, Legend, Tooltip } from 'recharts'
+import { ResponsiveContainer, XAxis, YAxis, BarChart, Bar, Legend, Tooltip } from 'recharts';
 
-import { formatMoney } from '../utils/formatMoney'
+import { formatMoney } from '../../utils/formatMoney'
 
-class DynamicChart extends Component {
+
+
+class DynamicBarChart extends Component {
   render() {
     const barArr = [{
       retireAmt: +this.props.retireAmt,
@@ -44,4 +46,4 @@ class DynamicChart extends Component {
   }
 }
 
-export default DynamicChart;
+export default DynamicBarChart;
