+++ src/index.js
@@ -1,19 +1,18 @@
-//React Imports
 import React from 'react';
 import { Router, Route, IndexRoute, browserHistory } from 'react-router';
 import { render } from 'react-dom';
 import { Provider } from 'react-redux';
-import store from './store.jsx';
-import Root from './components/Root';
-import CalculatorContainer from './containers/CalculatorContainer';
-import NotFound from './components/NotFound';
 
+import store from './store.jsx';
+
+import App from './elements/App';
+import CalculatorContainer from './containers/CalculatorContainer';
+import NotFound from './elements/NotFound';
 
 render(
   <Provider store={ store }>
     <Router history={ browserHistory } >
-      <Route path="/" component={ Root }>
-        <Route path="/calculator" component={ CalculatorContainer } />
+      <Route path="/" component={ App }>
         <IndexRoute component={ CalculatorContainer } />
       </Route>
       <Route path='*' component={ NotFound } />
